gg_GWAS_Summary

Creates a summary GWAS plot of significant associations.

Usage

gg_GWAS_Summary(
  folder = NULL,
  traits = list_Traits(),
  threshold = -log10(5e-08),
  sug.threshold = -log10(5e-06),
  models = c("MLM", "MLMM", "FarmCPU", "BLINK", "GLM"),
  colors = c("darkgreen", "darkred", "darkorange3", "steelblue", "darkgoldenrod2"),
  shapes = 21:25,
  hlines = NULL,
  vlines = NULL,
  vline.colors = rep("red", length(vlines)),
  vline.types = rep(1, length(vlines)),
  vline.legend = T,
  title = NULL,
  caption = paste0("Sig Threshold = ", threshold, " = Large\nSuggestive = ",
    sug.threshold, " = Small"),
  rowread = 2000,
  legend.position = "bottom",
  legend.rows = 1
)

Arguments

folder

Folder containing GWAS results.

traits

The traits to read.

threshold

Significant threshold.

sug.threshold

Suggestive threshold.

models

Models to read.

colors

Colors for each model.

shapes

The shape values to use for the different models. e.g., 21:25

hlines

Locations for horizontal lines. e.g., hlines = c(1.5,2.5).

vlines

Markers to be labelled with a vertical red line.

vline.colors

colors for each vertical line.

vline.types

lty for each vertical line.

vline.legend

Logical, display of vline color legend.

title

A title for the plot.

caption

A caption for the plot.

rowread

Number of rows to read for each GWAS results file.

legend.rows

Number of rows for the legend.

Value

A GWAS summary plot.